maps/ctmap: move CtEntry to types
